Make MoveOrDrop a trait
