+++ core/src/main/scala/Replay.scala
@@ -6,8 +6,6 @@ import chess.format.pgn.{ Parser, Reader, San, SanStr, Tag, Tags }
 import chess.format.{ Fen, Uci }
 import chess.variant.Variant
 
-import MoveOrDrop.*
-
 case class Replay(setup: Game, moves: List[MoveOrDrop], state: Game):
 
   lazy val chronoMoves: List[MoveOrDrop] = moves.reverse
