+++ test-kit/src/test/scala/DividerTest.scala
@@ -2,7 +2,6 @@ package chess
 
 import scala.language.implicitConversions
 
-import MoveOrDrop.*
 
 class DividerTest extends ChessTest:
