+++ test-kit/src/test/scala/perft/Perft.scala
@@ -4,7 +4,6 @@ package perft
 import chess.format.{ Fen, FullFen }
 import chess.variant.{ Chess960, Crazyhouse, Variant }
 
-import MoveOrDrop.*
 
 case class Perft(id: String, epd: FullFen, cases: List[TestCase]):
   import Perft.*
